+++ tests/Feature/Route/RouteCompletionContextTest.php
@@ -10,6 +10,7 @@ use Symfony\Lsp\Feature\DependencyInjection\DependencyInjectionSourceIndex;
 use Symfony\Lsp\Feature\Route\RouteCompletionContext;
 use Symfony\Lsp\Feature\Route\RouteReferenceExtractor;
 use Symfony\Lsp\Parser\Php\TolerantPhpParser;
+use Symfony\Lsp\Parser\QuotedArgumentMatcher;
 
 final class RouteCompletionContextTest extends TestCase
 {
@@ -21,7 +22,7 @@ final class RouteCompletionContextTest extends TestCase
         $php = str_replace('|', '', $php);
         $converter = new PositionConverter();
 
-        $extractor = new RouteReferenceExtractor($converter, new TolerantPhpParser(new Parser()));
+        $extractor = new RouteReferenceExtractor($converter, new TolerantPhpParser(new Parser()), new QuotedArgumentMatcher($converter));
         $context = RouteCompletionContext::fromPhp(
             $php,
             $converter->toPosition($php, $cursor),
@@ -98,7 +99,7 @@ final class RouteCompletionContextTest extends TestCase
         $cursor = strpos($php, "')");
         self::assertIsInt($cursor);
 
-        $extractor = new RouteReferenceExtractor($converter, new TolerantPhpParser(new Parser()));
+        $extractor = new RouteReferenceExtractor($converter, new TolerantPhpParser(new Parser()), new QuotedArgumentMatcher($converter));
         self::assertNull(RouteCompletionContext::fromPhp(
             $php,
             $converter->toPosition($php, $cursor),