+++ tests/Feature/Route/RouteCompletionHandlerTest.php
@@ -16,6 +16,7 @@ use Symfony\Lsp\Feature\Route\RouteCompletionHandler;
 use Symfony\Lsp\Feature\Route\RouteIndexRegistry;
 use Symfony\Lsp\Feature\Route\RouteReferenceExtractor;
 use Symfony\Lsp\Parser\Php\TolerantPhpParser;
+use Symfony\Lsp\Parser\QuotedArgumentMatcher;
 use Symfony\Lsp\Project\Project;
 use Symfony\Lsp\Project\ProjectRegistry;
 use Symfony\Lsp\Protocol\LspProtocolMapper;
@@ -241,7 +242,7 @@ final class RouteCompletionHandlerTest extends TestCase
             new LspProtocolMapper(),
             $indexes,
             $classIndexes ?? new DependencyInjectionSourceIndexRegistry(),
-            new RouteReferenceExtractor($converter, new TolerantPhpParser(new Parser())),
+            new RouteReferenceExtractor($converter, new TolerantPhpParser(new Parser()), new QuotedArgumentMatcher($converter)),
         );
     }
 }