+++ tests/Feature/Route/RouteDiagnosticPublisherTest.php
@@ -18,6 +18,7 @@ use Symfony\Lsp\Feature\Route\RouteIndexRegistry;
 use Symfony\Lsp\Feature\Route\RouteReferenceExtractor;
 use Symfony\Lsp\Feature\Route\TwigRouteReferenceExtractor;
 use Symfony\Lsp\Parser\Php\TolerantPhpParser;
+use Symfony\Lsp\Parser\QuotedArgumentMatcher;
 use Symfony\Lsp\Parser\TreeSitter\NativeTreeSitterParser;
 use Symfony\Lsp\Parser\TreeSitter\TreeSitterResultDecoder;
 use Symfony\Lsp\Parser\Twig\TwigCommentParser;
@@ -164,7 +165,7 @@ final class RouteDiagnosticPublisherTest extends TestCase
             $indexes->forProject($project)->replace(new Route('article_show', '/article/{id}', ['GET'], [], null, null));
             $converter = new PositionConverter();
             $classIndexes = new DependencyInjectionSourceIndexRegistry();
-            $phpExtractor = new RouteReferenceExtractor($converter, new TolerantPhpParser(new Parser()));
+            $phpExtractor = new RouteReferenceExtractor($converter, new TolerantPhpParser(new Parser()), new QuotedArgumentMatcher($converter));
             $twigExtractor = new TwigRouteReferenceExtractor($converter, new TwigDocumentParser(new NativeTreeSitterParser(new TreeSitterResultDecoder()), new TwigCommentParser()));
             $diagnosticProvider = new RouteDiagnosticPublisher(new DocumentContextResolver($documents, $projects), new LspProtocolMapper(), $indexes, $classIndexes, $phpExtractor, $twigExtractor);
             $diagnostics = $diagnosticProvider->diagnostics(['textDocument' => ['uri' => $uri]]);
@@ -255,7 +256,7 @@ final class RouteDiagnosticPublisherTest extends TestCase
                 new LspProtocolMapper(),
                 new RouteIndexRegistry(),
                 new DependencyInjectionSourceIndexRegistry(),
-                new RouteReferenceExtractor($positionConverter, new TolerantPhpParser(new Parser())),
+                new RouteReferenceExtractor($positionConverter, new TolerantPhpParser(new Parser()), new QuotedArgumentMatcher($positionConverter)),
                 new TwigRouteReferenceExtractor($positionConverter, new TwigDocumentParser(new NativeTreeSitterParser(new TreeSitterResultDecoder()), new TwigCommentParser())),
             )],
         );
@@ -331,7 +332,7 @@ final class RouteDiagnosticPublisherTest extends TestCase
                     new LspProtocolMapper(),
                     $routeIndexes,
                     new DependencyInjectionSourceIndexRegistry(),
-                    new RouteReferenceExtractor($positionConverter, new TolerantPhpParser(new Parser())),
+                    new RouteReferenceExtractor($positionConverter, new TolerantPhpParser(new Parser()), new QuotedArgumentMatcher($positionConverter)),
                     new TwigRouteReferenceExtractor($positionConverter, new TwigDocumentParser(new NativeTreeSitterParser(new TreeSitterResultDecoder()), new TwigCommentParser())),
                 )],
             ),